Corrected type of refineregionlifetime parameter

+++ src/enzo/StarParticleData.h
@@ -90,7 +90,7 @@ SPEXTERN float PopIIIColorMass;
 SPEXTERN int   PopIIIUseHypernova;
 SPEXTERN int   PopIIISupernovaExplosions;
 SPEXTERN int   PopIIIOutputOnFeedback;
-SPEXTERN int   PopIIIMustRefineRegionLifetime;
+SPEXTERN float   PopIIIMustRefineRegionLifetime;
 
 SPEXTERN int    StarClusterUseMetalField;
 SPEXTERN int    StarClusterHeliumIonization;
